consolidate tokensets api calles
