+++ src/utils/tokensetsApi.ts
@@ -1,3 +1,5 @@
+import { camelCase } from 'lodash'
+
 const baseURL = 'https://api.tokensets.com/public'
 
 export const fetchTokenBuySellData = (
@@ -13,5 +15,34 @@ export const fetchTokenBuySellData = (
   return fetch(requestUrl)
     .then((response) => response.json())
     .then((json) => json.buy_price || json.sell_price || {})
+    .catch((error) => console.log(error))
+}
+
+export const fetchSetComponents = (set: string) => {
+  const requestUrl = `${baseURL}/v2/portfolios/${set}`
+
+  return fetch(requestUrl)
+    .then((response) => response.json())
+    .then((response) => {
+      if (!response?.portfolio?.components) {
+        // undocumented API endpoint. Throw error if not expected response format
+        throw new Error('Invalid API response from Set Protocol service')
+      }
+      const {
+        portfolio: { components },
+      } = response
+      const formattedComponents = components.map((component: any) => {
+        const camelCasedComponent = Object.keys(component).reduce(
+          (comp: any, k: string) => ({
+            ...comp,
+            [camelCase(k)]: component[k],
+          }),
+          {}
+        )
+        return camelCasedComponent
+      })
+
+      return formattedComponents
+    })
     .catch((error) => console.log(error))
 }
